Moved Functions off MainScreen and Updated AccountSave
Changed:
- Moved Functions that are not connected to a visual on MainScreen
- Moved First time Run Check into FunctionStatus Class

Updated:
- Updated Account Save Class to save Raw Password in Account.ini

Removed:
- Properties File is now decommissioned
