+++ GameLauncher/app.config
@@ -2,9 +2,6 @@
 <configuration>
   <configSections>
     <section name="entityFramework" type="System.Data.Entity.Internal.ConfigFile.EntityFrameworkSection, EntityFramework, Version=6.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" requirePermission="false"/>
-    <sectionGroup name="userSettings" type="System.Configuration.UserSettingsGroup, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089">
-      <section name="GameLauncher.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" allowExeDefinition="MachineToLocalUser" requirePermission="false"/>
-    </sectionGroup>
   </configSections>
   <startup>
       <supportedRuntime version="v4.0.30319" sku=".NETFramework,Version=v4.8"/>
@@ -32,17 +29,4 @@
             </dependentAssembly>
        </assemblyBinding>
   </runtime>
-  <userSettings>
-    <GameLauncher.Properties.Settings>
-      <setting name="PasswordDecoded" serializeAs="String">
-        <value/>
-      </setting>
-      <setting name="IsRestarting" serializeAs="String">
-        <value>False</value>
-      </setting>
-      <setting name="IgnoreUpdateVersion" serializeAs="String">
-        <value/>
-      </setting>
-    </GameLauncher.Properties.Settings>
-  </userSettings>
 </configuration>
